Redirect to login page only when summiting a vote in detail page. Tests pass.
