+++ polls/tests.py
@@ -5,8 +5,11 @@ from unittest.mock import patch
 from django.test import TestCase
 from django.utils import timezone
 from django.urls import reverse
-
-from .models import Question
+import django.test
+from django.urls import reverse
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate # to "login" a user using code
+from .models import Question, Choice
 
 
 class QuestionModelTests(TestCase):
@@ -226,3 +229,87 @@ class QuestionResultViewTests(TestCase):
         url = reverse('polls:results', args=(past_question.id,))
         response = self.client.get(url)
         self.assertContains(response, past_question.question_text)
+
+
+class UserAuthTest(django.test.TestCase):
+
+    def setUp(self):
+        # superclass setUp creates a Client object and initializes test database
+        super().setUp()
+        self.username = "testuser"
+        self.password = "FatChance!"
+        self.user1 = User.objects.create_user(
+            username=self.username,
+            password=self.password,
+            email="testuser@nowhere.com"
+        )
+        self.user1.first_name = "Tester"
+        self.user1.save()
+        # we need a poll question to test voting
+        q = create_question("First Poll Question")
+        q.save()
+        # a few choices
+        for n in range(1, 4):
+            choice = Choice(choice_text=f"Choice {n}", question=q)
+            choice.save()
+        self.question = q
+
+    def test_logout(self):
+        """a user can logout using the logout url.
+
+        As an authenticated user,
+        when I visit /accounts/logout/
+        then I am logged out
+        and then redirected to the login page.
+        """
+        logout_url = reverse("logout")
+        # Authenticate the user.
+        # We want to logout this user, so we need to associate the
+        # user user with a session.  Setting client.user = ... doesn't work.
+        # Use Client.login(username, password) to do that.
+        # Client.login returns true on success
+        self.assertTrue(
+            self.client.login(username=self.username, password=self.password)
+        )
+        # visit the logout page
+        response = self.client.get(logout_url)
+        self.assertEqual(302, response.status_code)
+
+        # should redirect us to where? Polls index? Login?
+        self.assertRedirects(response, reverse('login'))
+
+    def test_login_view(self):
+        """a user can login using the login view."""
+        login_url = reverse("login")
+        # Can get the login page
+        response = self.client.get(login_url)
+        self.assertEqual(200, response.status_code)
+        # Can login using a POST request
+        # usage: client.post(url, {'key1":"value", "key2":"value"})
+        form_data = {"username": "testuser",
+                     "password": "FatChance!"
+                     }
+        response = self.client.post(login_url, form_data)
+        self.assertEqual(302, response.status_code)
+        # should redirect us to the polls index page ("polls:index")
+        self.assertRedirects(response, reverse("polls:index"))
+
+    def test_auth_required_to_vote(self):
+        """Authentication is required to submit a vote.
+
+        As an unauthenticated user,
+        when I submit a vote for a question,
+        then I am redirected to the login page
+          or I receive a 403 response (FORBIDDEN)
+        """
+        vote_url = reverse('polls:vote', args=[self.question.id])
+
+        # what choice to vote for?
+        choice = self.question.choice_set.first()
+        # the polls detail page has a form, each choice is identified by its id
+        form_data = {"choice": f"{choice.id}"}
+        response = self.client.post(vote_url, form_data)
+        # should be redirected to the login page
+        self.assertEqual(response.status_code, 302)  # could be 303
+        login_with_next = f"{reverse('login')}?next={vote_url}"
+        self.assertRedirects(response, login_with_next)
