+++ polls/views.py
@@ -5,7 +5,6 @@ from django.urls import reverse
 from django.views import generic
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.mixins import LoginRequiredMixin
 
 from .models import Question, Choice, Vote
 
@@ -35,7 +34,7 @@ class IndexView(generic.ListView):
         ).order_by('-pub_date')[:5]
 
 
-class DetailView(LoginRequiredMixin, generic.DetailView):
+class DetailView(generic.DetailView):
     """A class for detail view of a poll."""
     model = Question
     template_name = 'polls/detail.html'
@@ -46,10 +45,16 @@ class DetailView(LoginRequiredMixin, generic.DetailView):
             if not question.can_vote():
                 messages.error(request, "‼️ Voting is not allowed for this question.")
                 return HttpResponseRedirect(reverse('polls:index'))
-            return render(request, 'polls/detail.html', {
-                'question': question,
-                'voted_choice': question.get_voted_choice(request.user)
-            })
+            if request.user.is_authenticated:
+                return render(request, 'polls/detail.html', {
+                    'question': question,
+                    'voted_choice': question.get_voted_choice(request.user)
+                })
+            else:
+                messages.error(request, "🔒️ You haven't logged in. Please log in to vote.")
+                return render(request, 'polls/detail.html', {
+                    'question': question,
+                })
         except Question.DoesNotExist:
             messages.error(request, "‼️ The question you're looking for does not exist.")
             return HttpResponseRedirect(reverse('polls:index'))
@@ -73,8 +78,14 @@ def vote(request, question_id):
     question = get_object_or_404(Question, pk=question_id)
     try:
         selected_choice = question.choice_set.get(pk=request.POST['choice'])
-    except (KeyError, Choice.DoesNotExist):
+    except Choice.DoesNotExist:
         messages.error(request, "‼️ You didn't select a choice.")
+        return render(request, 'polls/detail.html', {
+            'question': question,
+            'voted_choice': question.get_voted_choice(request.user)
+        })
+    except KeyError:
+        messages.success(request, "🔑 You're logged in and able to vote.")
         return render(request, 'polls/detail.html', {
             'question': question,
             'voted_choice': question.get_voted_choice(request.user)
